+++ app/js/components/app.jsx
@@ -11,7 +11,6 @@ var Quickview = require('../components/quickview/index.jsx');
 var CenterProfileWindow = require('./profile/CenterProfileWindow.jsx');
 var CenterContactsWindow = require('./contacts/CenterContactsWindow.jsx');
 var SettingsWindow = require('ozp-react-commons/components/profile/SettingsWindow.jsx');
-
 var FeedbackModal = require('./management/user/FeedbackModal.jsx');
 var { ListingDeleteConfirmation } = require('./shared/DeleteConfirmation.jsx');
 
@@ -29,8 +28,7 @@ var App = React.createClass({
     },
 
     renderModal: function () {
-        var { listing, profile, contacts, settings, tab, action } = this.getQuery();
-
+        var { listing, profile, settings, contacts, tab, action} = this.getQuery();
         if (listing) {
             if (tab) {
                 var preview = action === 'preview';