+++ app/js/services/ListingService.js
@@ -6,6 +6,7 @@ var { PAGINATION_MAX } = require('ozp-react-commons/constants');
 var OzpAnalytics = require('../analytics/ozp-analytics');
 var ListingActions = require('../actions/ListingActions');
 var PaginatedListingsStore = require('../stores/PaginatedListingsStore');
+var UnpaginatedListingsStore = require('../stores/UnpaginatedListingsStore');
 
 function updateListingProperty(propName, value, listing) {
     var data = _.cloneDeep(listing);
@@ -38,9 +39,38 @@ ListingActions.fetchAllListings.listen(function (filter) {
         });
     }
 
+    var tmpPartial = _.partial(ListingActions.fetchAllListingsCompleted, _.clone(filter));
     ListingApi
         .getAllListings(nextLink, opts)
-        .then(_.partial(ListingActions.fetchAllListingsCompleted, filter));
+        .then(tmpPartial);
+});
+
+ListingActions.fetchAllListingsAtOnce.listen(function (filter) {
+    var unpaginatedList = UnpaginatedListingsStore.getListingsByFilter(filter),
+        opts = {},
+        nextLink;
+
+    if (unpaginatedList) {
+        unpaginatedList.expectPage();
+        nextLink = unpaginatedList.nextLink;
+    }
+    else {
+        // remove undefined values
+        _.forOwn(filter || {}, function (value, key) {
+            if(value !== null) {
+                opts[key] = value;
+            }
+        });
+
+        _.assign(opts, {
+            offset: 0
+        });
+    }
+
+    var tmpPartial = _.partial(ListingActions.fetchAllListingsAtOnceCompleted, _.clone(filter));
+    ListingApi
+        .getAllListings(nextLink, opts)
+        .then(tmpPartial);
 });
 
 ListingActions.fetchCounts.listen(function () {
@@ -93,7 +123,7 @@ ListingActions.fetchById.listen(function (id) {
     function processQuery(queryString) {
         var matches = queryString && queryString.match(/"[^"]*"|\S+/g),
             processedMatches = matches && matches.map(
-                m => /["\*]$/.test(m) ? m : m + '*'
+                m => /[""\*]$/.test(m) ? m : m + '*'   
             );
 
         return processedMatches && processedMatches.join(' ');
@@ -143,7 +173,7 @@ ListingActions.saveReview.listen(function (listing, review) {
             ListingActions.fetchById(listing.id);
             ListingActions.fetchReviews(listing);
             ListingActions.saveReviewCompleted(listing, response);
-            OzpAnalytics.trackListingReview(listing.title);
+            OzpAnalytics.trackListingReview(listing.title, listing.agencyShort);
         })
         .fail(ListingActions.saveReviewFailed);
 });
@@ -166,7 +196,7 @@ ListingActions.launch.listen(function (listing) {
 ListingActions.save.listen(function (data) {
     var isNew = !data.id;
 
-    if (isNew) { OzpAnalytics.trackListingCreation(data.title); }
+    if (isNew) { OzpAnalytics.trackListingCreation(data.title, data.agency); }
 
     if (/\s/g.test(data.launchUrl)) { data.launchUrl = data.launchUrl.replace(/ /g,"%20"); }
 
@@ -187,12 +217,12 @@ ListingActions.enable.listen(setEnabled.bind(null, true));
 ListingActions.disable.listen(setEnabled.bind(null, false));
 
 ListingActions.approve.listen(function (listing) {
-    OzpAnalytics.trackListingApproval(listing.title);
+    OzpAnalytics.trackListingApproval(listing.title, listing.agencyShort);
     updateListingProperty('approvalStatus', 'APPROVED', listing);
 });
 
 ListingActions.approveByOrg.listen(function (listing) {
-    OzpAnalytics.trackListingOrgApproval(listing.title);
+    OzpAnalytics.trackListingOrgApproval(listing.title, listing.agencyShort);
     updateListingProperty('approvalStatus', 'APPROVED_ORG', listing);
 });